Switch to interrupt driven drawing
This change is pretty disgusting but an interrupt > 50uS is likely to
cause the WiFi or TCP connection to drop. I think this is due to the
WiFi connection on an Uno R4 being driven over UART, which itself
requires interrupts. Anyway, breaking this into smaller segments seems
to avoid it crashing, and still allows for a framerate of > 150fps,
without the occasionally blip caused by mqtt.loop().

Yet to be tested on an actual departure board.